All players added to database

+++ acquire_data.py
@@ -1,39 +1,20 @@
-import requests
-import csv,json
+import requests, csv,json, os.path
 
 user_agent = {'User-Agent': 'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_11_6) '
                             'AppleWebKit/537.36 (KHTML, like Gecko) Chrome/58.0.3029.110 '
                             'Safari/537.36'}
 
-
 def upload_from_json(filename):
-    with open('shot_charts/{}.json'.format(filename), 'r') as fp:
+    with open('{}.json'.format(filename), 'r') as fp:
         data = json.load(fp)
         return data
 
 def save_json(filename, data):
-    with open('shot_charts/{}.json'.format(filename), 'w') as fp:
-        json.dump(data, fp)
-
-def get_player_ids(user_agent):
-    players_url = "http://stats.nba.com/stats/leaguedashplayerstats?College=&Conference=&" \
-                  "Country=&DateFrom=&DateTo=&Division=&DraftPick=&DraftYear=&GameScope=&" \
-                  "GameSegment=&Height=&LastNGames=0&LeagueID=00&Location=&MeasureType=Base" \
-                  "&Month=0&OpponentTeamID=0&Outcome=&PORound=0&PaceAdjust=N&PerMode=PerGame" \
-                  "&Period=0&PlayerExperience=&PlayerPosition=&PlusMinus=N&Rank=N&Season=" \
-                  "2016-17&SeasonSegment=&SeasonType=Playoffs&ShotClockRange=&StarterBench=&" \
-                  "TeamID=0&VsConference=&VsDivision=&Weight="
-
-    response = requests.get(players_url, headers = user_agent)
-    response.raise_for_status()
-    data = response.json()['resultSets'][0]
-    player_stats = data['rowSet']
-    player_ids = {}
-
-    for player in player_stats:
-        player_id = player[0]
-        player_ids[player_id] = player[1]
-    save_json("player_ids", player_ids)
+    if os.path.isfile("{}.json".format(filename)):
+        pass
+    else:
+        with open('{}.json'.format(filename), 'w') as fp:
+            json.dump(data, fp)
 
 def get_player_shots(player_id, user_agent):
     shots_url = "http://stats.nba.com/stats/shotchartdetail?AheadBehind=&CFID=33&" \
@@ -51,12 +32,8 @@ def get_player_shots(player_id, user_agent):
     response = requests.get(shots_url, headers = user_agent)
     response.raise_for_status()
     data = response.json()['resultSets'][0]
-    save_json(player_id, data)
+    return data
 
-
-def get_shot_charts(player_ids):
-    for player_id in player_ids:
-        get_player_shots(player_id, user_agent)
 
 def write_shots_to_csv(filename, shot_headers, shot_charts):
     shot_charts.append(shot_headers)
@@ -67,7 +44,7 @@ def write_shots_to_csv(filename, shot_headers, shot_charts):
 def get_target_shot(player_id, shot_id):
     target_shot = []
     headers = []
-    with open('shot_charts/{}.json'.format(player_id), 'r') as fp:
+    with open('shotCharts/shot_charts/{}.json'.format(player_id), 'r') as fp:
         data = json.load(fp)
         print data.keys()
         shots = data['rowSet']
@@ -93,10 +70,37 @@ def get_file_json(filename):
         return data
 
 def save_player_images():
-    player_ids = get_file_json("player_ids")
+    player_ids = get_file_json("shotCharts/player_ids")
     for player in player_ids:
         url = "http://stats.nba.com/media/players/230x185/{}.png".format(player)
         response = requests.get(url)
         if response.status_code == 200:
-            with open("static/images/{}.png".format(player), 'wb') as f:
-                f.write(response.content)
+            filepath = "shotCharts/static/images/{}.png".format(player)
+            if not os.path.isfile(filepath):
+                with open(filepath, 'wb') as f:
+                    f.write(response.content)
+
+def save_player_ids_web(user_agent):
+    players_url = "http://stats.nba.com/stats/leaguedashplayerstats?College=&Conference=&Country=" \
+                  "&DateFrom=&DateTo=&Division=&DraftPick=&DraftYear=&GameScope=&GameSegment=&H" \
+                  "eight=&LastNGames=0&LeagueID=00&Location=&MeasureType=Base&Month=0&OpponentTeamID" \
+                  "=0&Outcome=&PORound=0&PaceAdjust=N&PerMode=PerGame&Period=0&PlayerExperience=&" \
+                  "PlayerPosition=&PlusMinus=N&Rank=N&Season=2016-17&SeasonSegment=&SeasonType=Regular" \
+                  "+Season&ShotClockRange=&StarterBench=&TeamID=0&VsConference=&VsDivision=&Weight="
+
+    response = requests.get(players_url, headers = user_agent)
+    response.raise_for_status()
+    data = response.json()['resultSets'][0]
+    player_stats = data['rowSet']
+    player_ids = {}
+
+    for player in player_stats:
+        player_id = player[0]
+        player_ids[player_id] = player[1]
+    save_json("player_ids", player_ids)
+
+def get_all_shot_charts():
+    player_ids = get_file_json("player_ids")
+    for player_id in player_ids:
+        data = get_player_shots(player_id, user_agent)
+        save_json("shotCharts/shot_charts/{}".format(player_id), data)
